+++ dailyWeb/src/main/java/cn/wenzhuo4657/dailyWeb/infrastructure/database/repository/TypesRepository.java
@@ -39,7 +39,7 @@ public class TypesRepository implements ITypesRepository {
     }
 
     @Override
-    public List<TypeDto> getAllByUserId(Integer userId) {
+    public List<TypeDto> getAllByUserId(Long userId) {
         log.warn("暂时没有处理动态路由----");
 //        todo 这里需要调整为动态路由，根据用户身份来做权限控制
         List<DocsType> all = docsTypeDao.getAll();
@@ -54,7 +54,7 @@ public class TypesRepository implements ITypesRepository {
     }
 
     @Override
-    public List<Docs> getDocsIdByTypeId(Integer userId, Integer typeId) {
+    public List<Docs> getDocsIdByTypeId(Long userId, Integer typeId) {
         List<Docs> list = userAuthDao.queryByUserIdAndtypeId(userId, typeId);
         return list;
     }
